backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
aiomysql==0.2.0
aiosqlite==0.20.0
pydantic==2.6.4
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.3
aiohttp==3.9.3
httpx==0.27.0
pytest==8.1.1
pytest-asyncio==0.23.6
